Move subposition to seperate data file
